requirements.txt
pydoll-python
pandas
beautifulsoup4
